fix: dark & light mode refresh flickering fix
