+++ RtM/RtM/Components/Layout/MainLayout.razor
@@ -1,13 +1,16 @@
 ﻿@inherits LayoutComponentBase
 
-<div class="h-screen flex flex-col bg-slate-200 dark:bg-gray-800 text-slate-900 dark:text-slate-300">
-    <NavMenu />
-    <div class="flex-grow container mx-auto flex flex-col text-slate-900 dark:text-slate-300">
-        <main class="flex-grow w-full bg-slate-200 dark:bg-gray-800 p-3 pb-0">
-            <div class="rounded-md bg-slate-300 dark:bg-gray-900 p-4 w-full h-full">
-                @Body
-            </div>
-        </main>
-        <Footer />
+@if (theme.HasValue)
+{
+    <div class="h-screen flex flex-col bg-slate-200 dark:bg-gray-800 text-slate-900 dark:text-slate-300">
+        <NavMenu theme="theme" ChangeTheme="ChangeTheme" />
+        <div class="flex-grow container mx-auto flex flex-col text-slate-900 dark:text-slate-300">
+            <main class="flex-grow w-full bg-slate-200 dark:bg-gray-800 p-3 pb-0">
+                <div class="rounded-md bg-slate-300 dark:bg-gray-900 p-4 w-full h-full">
+                    @Body
+                </div>
+            </main>
+            <Footer />
+        </div>
     </div>
-</div>
+}
