+++ RtM/RtM/Components/Layout/NavMenu.razor
@@ -1,16 +1,17 @@
 ﻿@using Blazor.Heroicons.Solid
 @inherits LayoutComponentBase
 
-<header class="w-full bg-slate-300 dark:bg-slate-900">
-    <div class="container mx-auto p-2 flex items-center justify-between">
-        <div>
-            <NavLink href="/" class="hover:text-slate-700 dark:hover:text-slate-400 px-2 py-2 text-sm font-medium">Home</NavLink>
-            <NavLink href="/about" class="hover:text-slate-700 dark:hover:text-slate-400 px-2 py-2 text-sm font-medium">About</NavLink>
-        </div>
-        <div class="mr-2">
-            <button @onclick="ChangeTheme" class="bg-slate-400 dark:bg-slate-600 rounded-md hover:bg-slate-500 dark:hover:bg-slate-500 px-1 py-1">
-                @{
-                    if (toggleDark)
+@if (Theme.HasValue)
+{
+    <header class="w-full bg-slate-300 dark:bg-slate-900">
+        <div class="container mx-auto p-2 flex items-center justify-between">
+            <div>
+                <NavLink href="/" class="hover:text-slate-700 dark:hover:text-slate-400 px-2 py-2 text-sm font-medium">Home</NavLink>
+                <NavLink href="/about" class="hover:text-slate-700 dark:hover:text-slate-400 px-2 py-2 text-sm font-medium">About</NavLink>
+            </div>
+            <div class="mr-2">
+                <button @onclick="ChangeTheme" class="bg-slate-400 dark:bg-slate-600 rounded-md hover:bg-slate-500 dark:hover:bg-slate-500 px-1 py-1">
+                    @if (Theme.Value)
                     {
                         <SunIcon class="h-6 w-6 text-slate-300" />
                     }
@@ -18,8 +19,8 @@
                     {
                         <MoonIcon class="h-6 w-6 text-slate-300" />
                     }
-                }
-            </button>
+                </button>
+            </div>
         </div>
-    </div>
-</header>
+    </header>
+}
